# docs/openapi.yaml
openapi: 3.0.0
info:
  title: API MontsSystem
  version: 1.0.0
  description: API para gestão de vendas e clientes
servers:
  - url: http://localhost:3001/api
    description: Servidor de desenvolvimento
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Sale:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: ID único da venda
        clientId:
          type: string
          description: ID do cliente
          nullable: true
        clientName:
          type: string
          description: Nome do cliente
        phone:
          type: string
          description: Telefone do cliente
        cpfOrCnpj:
          type: string
          description: CPF ou CNPJ do cliente
          nullable: true
        address:
          type: string
          description: Endereço do cliente
          nullable: true
        saleDate:
          type: string
          format: date
          description: Data da venda
          nullable: true
        warrantyEndDate:
          type: string
          format: date
          description: Data de fim da garantia
          nullable: true
        productDescription:
          type: string
          description: Descrição do produto/serviço
          nullable: true
        saleValue:
          type: number
          format: float
          description: Valor da venda em reais
          nullable: true
          minimum: 0
        saleValueCents:
          type: integer
          description: Valor da venda em centavos
          nullable: true
          minimum: 0
        isMonthly:
          type: boolean
          description: Indica se é uma venda mensal/assinatura
          default: false
        entryValue:
          type: number
          format: float
          description: Valor de entrada para clientes mensais
          nullable: true
          minimum: 0
        monthlyValue:
          type: number
          format: float
          description: Valor mensal para clientes de assinatura
          nullable: true
          minimum: 0
        nextPaymentDate:
          type: string
          format: date
          description: Data do próximo pagamento
          nullable: true
        contractFile:
          type: string
          description: Caminho do arquivo de contrato
          nullable: true
        invoiceFile:
          type: string
          description: Caminho do arquivo de nota fiscal
          nullable: true
      required:
        - clientName
        - phone
        - isMonthly
    CreateSaleRequest:
      type: object
      properties:
        clientId:
          type: string
          description: ID do cliente
        clientName:
          type: string
          description: Nome do cliente
        phone:
          type: string
          description: Telefone do cliente
        cpfOrCnpj:
          type: string
          description: CPF ou CNPJ do cliente
        address:
          type: string
          description: Endereço do cliente
        saleDate:
          type: string
          format: date
          description: Data da venda
        warrantyEndDate:
          type: string
          format: date
          description: Data de fim da garantia
        productDescription:
          type: string
          description: Descrição do produto/serviço
        saleValue:
          oneOf:
            - type: number
              format: float
            - type: string
          description: Valor da venda (aceita "R$ 1.234,56" ou número)
          example: "R$ 1.234,56"
        isMonthly:
          type: boolean
          description: Indica se é uma venda mensal/assinatura
          default: false
        entryValue:
          oneOf:
            - type: number
              format: float
            - type: string
          description: Valor de entrada (aceita formatos monetários)
          example: "200,00"
        monthlyValue:
          oneOf:
            - type: number
              format: float
            - type: string
          description: Valor mensal (aceita formatos monetários)
          example: "150,00"
        nextPaymentDate:
          type: string
          format: date
          description: Data do próximo pagamento
      required:
        - clientName
        - phone
    Ticket:
      type: object
      properties:
        id:
          type: string
          format: uuid
        clientId:
          type: string
          format: uuid
          description: ID do cliente vinculado
        clientName:
          type: string
          nullable: true
          description: Nome do cliente (denormalizado na resposta)
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [open, in_progress, waiting_customer, resolved, cancelled]
          default: open
        priority:
          type: string
          enum: [low, medium, high, urgent]
          default: medium
        dueDate:
          type: string
          format: date-time
          nullable: true
        assigneeName:
          type: string
          nullable: true
        attachments:
          type: array
          items:
            type: string
            format: uri
          description: URLs públicas dos anexos no DigitalOcean Spaces
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - clientId
        - title
        - description
        - status
        - priority
        - createdAt
        - updatedAt
    TicketUpdate:
      type: object
      properties:
        clientId:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [open, in_progress, waiting_customer, resolved, cancelled]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        dueDate:
          type: string
          format: date-time
        assigneeName:
          type: string
        replaceAttachments:
          type: boolean
          description: Quando true, substitui os anexos existentes pelos novos enviados
        files:
          type: array
          items:
            type: string
            format: binary
...
    UserPublic:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        name:
          type: string
        email:
          type: string
          nullable: true
        role:
          type: string
          enum: [admin, user]
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - username
        - name
        - role
        - active
    UserDetail:
      allOf:
        - $ref: '#/components/schemas/UserPublic'
        - type: object
          properties:
            permissions:
              type: array
              items:
                $ref: '#/components/schemas/PermissionItem'
    PermissionItem:
      type: object
      properties:
        resource:
          type: string
          enum: [customers, sales, invoices, boletos, mensalidades, nfse, whatsapp, tickets, company, integrations, users]
        level:
          type: string
          enum: [view, edit]
      required:
        - resource
        - level
...
    CreateUserRequest:
      type: object
      required:
        - username
        - password
        - name
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 64
        password:
          type: string
          minLength: 6
        name:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [admin, user]
          default: user
        active:
          type: boolean
          default: true
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/PermissionItem'
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
          nullable: true
        password:
          type: string
          minLength: 6
          description: Omitir para não alterar
        role:
          type: string
          enum: [admin, user]
        active:
          type: boolean
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/PermissionItem'
paths:
...
  /users:
    get:
      summary: Listar todos os usuários (apenas admin)
      tags:
        - Users
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de usuários
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserPublic'
        '403':
          description: Sem permissão

    post:
      summary: Criar usuário (apenas admin)
      tags:
        - Users
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: Usuário criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPublic'
        '400':
          description: Erro de validação
        '403':
          description: Sem permissão
        '409':
          description: Username já em uso
  /users/{id}:
    get:
      summary: Buscar usuário por id (apenas admin)
      tags:
        - Users
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Usuário encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetail'
        '403':
          description: Sem permissão
        '404':
          description: Não encontrado

    put:
      summary: Atualizar usuário (apenas admin)
      description: |
        Campos vazios são ignorados (partial update). Mudança de senha,
        role ou active incrementa `tokenVersion`, invalidando todas as
        sessões existentes do usuário.
      tags:
        - Users
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: Atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPublic'
        '400':
          description: |
            Erro de validação ou tentativa de despromover/desativar/deletar
            o último administrador ativo do sistema.
        '403':
          description: Sem permissão
        '404':
          description: Não encontrado

    delete:
      summary: Excluir usuário (apenas admin)
      tags:
        - Users
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Excluído
        '400':
          description: Tentativa de excluir o último admin
        '403':
          description: Sem permissão
        '404':
          description: Não encontrado
  /sales:
    get:
      summary: Listar todas as vendas
      tags:
        - Sales
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de vendas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Sale'
              examples:
                vendas_exemplo:
                  value:
                    - id: "abc123"
                      clientId: "cli-1"
                      clientName: "João Silva"
                      phone: "11999999999"
                      saleValue: 1234.56
                      saleValueCents: 123456
                      saleDate: "2025-10-01"
                      isMonthly: false
                      productDescription: "Instalação X"
                    - id: "def456"
                      clientId: "cli-2"
                      clientName: "Maria Santos"
                      phone: "11888888888"
                      isMonthly: true
                      entryValue: 200.00
                      monthlyValue: 150.00
                      nextPaymentDate: "2025-11-01"
                      productDescription: "Contrato mensal Y"

    post:
      summary: Criar nova venda
      tags:
        - Sales
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSaleRequest'
            examples:
              venda_unica:
                summary: Venda única
                value:
                  clientId: "cli-1"
                  clientName: "João Silva"
                  phone: "11999999999"
                  saleValue: "R$ 1.234,56"
                  saleDate: "2025-10-01"
                  productDescription: "Instalação X"
                  isMonthly: false
              cliente_mensal:
                summary: Cliente mensal
                value:
                  clientId: "cli-2"
                  clientName: "Maria Santos"
                  phone: "11888888888"
                  isMonthly: true
                  entryValue: "200,00"
                  monthlyValue: "150,00"
                  nextPaymentDate: "2025-11-01"
                  productDescription: "Contrato mensal Y"
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CreateSaleRequest'
      responses:
        '201':
          description: Venda criada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sale'
        '400':
          description: Dados inválidos
        '401':
          description: Token de autorização inválido
  /sales/{id}:
    get:
      summary: Obter venda por ID
      tags:
        - Sales
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Dados da venda
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sale'
        '404':
          description: Venda não encontrada

    put:
      summary: Atualizar venda
      tags:
        - Sales
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSaleRequest'
      responses:
        '200':
          description: Venda atualizada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sale'
        '404':
          description: Venda não encontrada

    delete:
      summary: Remover venda
      tags:
        - Sales
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Venda removida com sucesso
        '404':
          description: Venda não encontrada
...
  /tickets:
    get:
      summary: Listar todos os chamados
      tags:
        - Tickets
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [open, in_progress, waiting_customer, resolved, cancelled]
          description: Filtrar por status
        - name: clientId
          in: query
          required: false
          schema:
            type: string
            format: uuid
          description: Filtrar por cliente
        - name: priority
          in: query
          required: false
          schema:
            type: string
            enum: [low, medium, high, urgent]
          description: Filtrar por prioridade
      responses:
        '200':
          description: Lista de chamados
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Ticket'
    post:
      summary: Criar novo chamado
      tags:
        - Tickets
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                clientId:
                  type: string
                  format: uuid
                title:
                  type: string
                description:
                  type: string
                status:
                  type: string
                  enum: [open, in_progress, waiting_customer, resolved, cancelled]
                priority:
                  type: string
                  enum: [low, medium, high, urgent]
                dueDate:
                  type: string
                  format: date-time
                assigneeName:
                  type: string
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: Anexos do chamado (PDF/JPG/PNG)
              required:
                - clientId
                - title
                - description
      responses:
        '201':
          description: Chamado criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ticket'
        '404':
          description: Cliente não encontrado
  /tickets/{id}:
    get:
      summary: Obter um chamado pelo ID
      tags:
        - Tickets
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Chamado encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ticket'
        '404':
          description: Chamado não encontrado
    put:
      summary: Atualizar um chamado
      tags:
        - Tickets
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/TicketUpdate'
      responses:
        '200':
          description: Chamado atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ticket'
        '404':
          description: Chamado ou cliente não encontrado
    delete:
      summary: Remover um chamado
      tags:
        - Tickets
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Chamado removido
        '404':
          description: Chamado não encontrado